+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing so far
+### Fixed
+- Fix error when cancelling game selection
 
 ## [1.0.1] - 2019-11-25
 ### Fixed