+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1]
 ### Added
 - [Playnite #10](https://playnite.link/forum/thread-24-post-67.html#pid67) Add option to enable incomplete dates
 ### Changed
@@ -43,7 +43,8 @@ Special Thanks to [darklinkpower](https://github.com/darklinkpower) for testing 
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Mithnar/PlayniteVndb/compare/1.0.1...HEAD
+[Unreleased]: https://github.com/Mithnar/PlayniteVndb/compare/1.1...HEAD
+[1.1]: https://github.com/Mithnar/PlayniteVndb/compare/1.0.1...1.1
 [1.0.1]: https://github.com/Mithnar/PlayniteVndb/compare/1.0...1.0.1
 [1.0]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC3...1.0
 [1.0-RC3]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC2...1.0-RC3